add cypress feature testing
