+++ bin/test
@@ -4,11 +4,13 @@ set -e
 
 bin/checkinputs "$@"
 
+DIR=$(dirname "$0")
+
 export OX_PROJECT=$1
 echo Test: $OX_PROJECT
-args=${@:2}
-if [[ ! $args ]]; then
-   args='--watch'
-fi
 
-$(npm bin)/jest --verbose false --config ./configs/test/jest.$OX_PROJECT.js $args
+if [[ $2 = "integration" ]]; then
+  $DIR/test-integration $OX_PROJECT ${@:3}
+else
+  $DIR/test-unit $OX_PROJECT ${@:2}
+fi
