+++ package.json
@@ -13,9 +13,11 @@
   },
   "scripts": {
     "preci": "npm run validate",
-    "ci": "./bin/test-ci",
     "test": "./bin/test",
     "tdd": "./bin/tdd",
+    "tutor:test:server": "node tutor/scripts/test-server/index.js",
+    "tutor:test:cypress": "./bin/test tutor integration",
+    "tutor:test:integration": "concurrently -n tutor:test:cypress tutor:test:server",
     "precoverage": "npm run validate",
     "serve": "./bin/serve",
     "cached-serve": "CACHE=true npm run serve",
@@ -58,6 +60,7 @@
     "big.js": "^5.2.2",
     "bootstrap": "^4.3.1",
     "classnames": "2.2.6",
+    "concurrently": "^5.0.2",
     "core-decorators": "0.20.0",
     "core-js": "^3.2.1",
     "create-react-class": "^15.6.3",
@@ -128,6 +131,7 @@
     "twix": "1.3.0",
     "underscore": "1.9.1",
     "validator": "^11.1.0",
+    "wait-on": "^3.3.0",
     "weak-map": "^1.0.5",
     "when-dom-ready": "^1.2.12",
     "youtube-player": "https://github.com/nathanstitt/youtube-player.git#iframe-or-window"
@@ -150,6 +154,7 @@
     "chrome-launcher": "^0.11.2",
     "coveralls": "3.0.6",
     "css-loader": "3.2.0",
+    "cypress": "^3.8.1",
     "enzyme": "^3.10.0",
     "eslint": "5.16.0",
     "eslint-plugin-react": "7.14.3",
