+++ tutor/specs/integration/dashboard.spec.js
@@ -1,0 +1,10 @@
+context('Dashboard', () => {
+  beforeEach(() => {
+    cy.visit('/dashboard');
+  });
+
+  it('hides value prop and displays cards', () => {
+    cy.getTestElement('value-prop').find('button').click();
+    cy.get('[data-is-preview=false][data-course-id]');
+  });
+});
